refactor: organize grid protocol package
