+++ dev/examples/01_grid_tutorial.py
@@ -1,18 +1,6 @@
 # SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES.
 # SPDX-FileCopyrightText: All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
 
 # %%
 """
@@ -21,13 +9,8 @@ Earth2Studio Grids
 
 Define, register, infer, and select spatial grid geometry.
 
-A grid describes spatial geometry without containing weather data. In this tutorial
-you will learn:
-
-- What the structural grid interface requires
-- How to resolve and register grid definitions
-- How to infer grids from Xarray coordinates
-- How grid definitions support selection and future regridding
+A grid describes geometry without containing weather data. This tutorial shows the
+structural interface, registry, Xarray coordinates, HEALPix layouts, and selection.
 """
 
 # /// script
@@ -39,13 +22,11 @@ you will learn:
 # %%
 # Define a Grid Object
 # --------------------
-# A grid object structurally implements
-# [`GridDefinition`][earth2studio.utils.grid.GridDefinition]. It does not inherit from
-# a grid base class. The protocol defines dimension order, shape, topology, coordinate
-# generation, metadata, selection, and stable identity.
+# A grid object implements [`GridDefinition`][earth2studio.grids.GridDefinition]
+# structurally. It does not inherit from a grid base class.
 
 # %%
-# .. literalinclude:: ../../earth2studio/utils/grid.py
+# .. literalinclude:: ../../earth2studio/grids/base.py
 #    :language: python
 #    :start-after: # sphinx - grid protocol start
 #    :end-before: # sphinx - grid protocol end
@@ -54,33 +35,26 @@ you will learn:
 import numpy as np
 import xarray as xr
 
-import earth2studio.utils.grid as e2s
+import earth2studio.grids as grids
 
 # %%
 # Resolve a Known Grid
 # --------------------
-# The registry maps stable names and aliases to complete grid definitions.
+# The registry maps stable names and aliases to complete definitions. Metadata is a
+# serializable summary suitable for Xarray attributes and external tools.
 
 # %%
-hrrr = e2s.resolve_grid("hrrr")
-crs = hrrr.crs
-grid_summary = {
-    "type": type(hrrr).__name__,
-    "dims": hrrr.dims,
-    "shape": hrrr.shape,
-    "topology": hrrr.topology,
-    "crs": crs.name if crs is not None else None,
-}
-print(grid_summary)
+hrrr = grids.resolve_grid("hrrr")
+print(hrrr.to_metadata())
 
 # %%
 # Register a Projected Grid
 # -------------------------
 # A projected grid requires ordered ``y`` and ``x`` coordinates plus any CRS accepted
-# by PyProj. Registration gives the complete definition a stable identity.
+# by PyProj.
 
 # %%
-regional = e2s.ProjectedGrid(
+regional = grids.ProjectedGrid(
     y=np.arange(3) * 3_000.0,
     x=np.arange(4) * 3_000.0,
     coordinate_reference_system=(
@@ -88,30 +62,47 @@ regional = e2s.ProjectedGrid(
         "+datum=WGS84 +units=m +type=crs"
     ),
 )
-e2s.register_grid("tutorial-lcc", regional)
+grids.register_grid("tutorial-lcc", regional)
 
-indexes = regional.coordinates(only_index=True)
-coordinates = regional.coordinates()
+indexes = regional.coords(only_index=True)
+coordinates = regional.coords()
 print(regional.shape, coordinates["lat"].shape)
 
 # %%
 # Populate Xarray Coordinates
 # ---------------------------
-# Dimension coordinates preserve the grid's spatial order. Latitude and longitude
-# are auxiliary coordinates, so projected ``y`` and ``x`` remain the array dimensions.
+# Dimension coordinates preserve spatial order. Latitude and longitude remain
+# auxiliary coordinates, so projected ``y`` and ``x`` are the array dimensions.
 
 # %%
 geometry = xr.Dataset(coords=coordinates)
 print(geometry)
 
-# Field data can later reuse these coordinates with
+# Field data can reuse these coordinates with
 # ``xr.DataArray(data, dims=regional.dims, coords=geometry.coords)``.
+
+# %%
+# Represent HEALPix Layouts
+# -------------------------
+# Ordering and layout are separate. DLESyM uses Earth2Grid's north-origin,
+# clockwise XY convention with explicit face, height, and width dimensions.
+
+# %%
+dlesym = grids.HEALPixGrid(
+    level=2,
+    ordering="xy",
+    layout="face",
+    xy_origin="north",
+    xy_clockwise=True,
+)
+print(dlesym.to_metadata())
+print(dlesym.coords(only_index=True))
 
 # %%
 # Infer Xarray Grids
 # ------------------
-# Registration is optional. Independent one-dimensional latitude and longitude
-# coordinates define a rectilinear grid.
+# Independent one-dimensional latitude and longitude coordinates define a
+# rectilinear grid.
 
 # %%
 latlon = xr.DataArray(
@@ -119,11 +110,10 @@ latlon = xr.DataArray(
     dims=("lat", "lon"),
     coords={"lat": [40.0, 39.0], "lon": [250.0, 251.0, 252.0]},
 )
-print(type(e2s.infer_grid(latlon)).__name__)
+print(type(grids.infer_grid(latlon)).__name__)
 
 # %%
-# Arbitrary locations use ``x`` as the ordered index with auxiliary latitude and
-# longitude coordinates.
+# Arbitrary locations use ``x`` as their ordered index.
 
 # %%
 points = xr.Dataset(
@@ -133,20 +123,16 @@ points = xr.Dataset(
         "lon": ("x", [-97.4, -74.0, 0.1]),
     }
 )
-point_grid = e2s.infer_grid(points)
-print(point_grid.dims, point_grid.topology, point_grid.coordinates(only_index=True))
+point_grid = grids.infer_grid(points)
+print(point_grid.dims, point_grid.topology, point_grid.coords(only_index=True))
 
 # %%
 # Select a Subdomain
 # ------------------
-# Geographic bounds are translated into normal Xarray indexers. Field data remains
-# outside the grid implementation.
+# Geographic bounds become normal Xarray indexers. Field data remains outside the
+# grid implementation.
 
 # %%
-grid = e2s.infer_grid(latlon)
-indexers = grid.subset_indexers(
-    latlon.coords,
-    bounds=(-110, 38, -90, 41),
-)
-subset = latlon.isel(indexers)
-print(subset)
+grid = grids.infer_grid(latlon)
+indexers = grid.subset_indexers(latlon.coords, bounds=(-110, 38, -90, 41))
+print(latlon.isel(indexers))
